Рефакторинг диалога

+++ src/main/kotlin/su/vshk/billing/bot/service/processor/ProcessorService.kt
@@ -7,13 +7,12 @@ import reactor.util.context.ContextView
 import su.vshk.billing.bot.dao.model.Command
 import su.vshk.billing.bot.dao.model.GenericCommand
 import su.vshk.billing.bot.dao.model.UserEntity
+import su.vshk.billing.bot.dao.service.UserDaoService
 import su.vshk.billing.bot.dialog.DialogProcessor
+import su.vshk.billing.bot.message.ResponseMessageService
 import su.vshk.billing.bot.message.dto.RequestMessageItem
 import su.vshk.billing.bot.message.dto.ResponseMessageItem
-import su.vshk.billing.bot.message.ResponseMessageService
-import su.vshk.billing.bot.service.LoginMessageService
 import su.vshk.billing.bot.service.PaymentNotificationService
-import su.vshk.billing.bot.dao.service.UserDaoService
 import su.vshk.billing.bot.service.executor.CommandExecutor
 import su.vshk.billing.bot.util.errorTraceId
 import su.vshk.billing.bot.util.getLogger
@@ -22,7 +21,6 @@ import su.vshk.billing.bot.util.infoTraceId
 @Service
 class ProcessorService(
     private val dialogProcessor: DialogProcessor,
-    private val loginMessageService: LoginMessageService,
     private val responseMessageService: ResponseMessageService,
     private val executors: List<CommandExecutor>,
     private val paymentNotificationService: PaymentNotificationService,
@@ -70,22 +68,16 @@ class ProcessorService(
     }
 
     private fun updateDialog(request: RequestMessageItem): Mono<ResponseMessageItem> =
-        when {
-            dialogProcessor.isLoginDialog(request.chatId) ->
-                loginMessageService.add(telegramId = request.chatId, messageId = request.messageId)
-                    .flatMap { doUpdateDialog(chatId = request.chatId, option = request.input) }
-
-            request.isTextUpdate ->
-                responseMessageService.deleteMessage(request.messageId).toMono()
-
-            else -> doUpdateDialog(chatId = request.chatId, option = request.input)
+        if (request.isTextUpdate && dialogProcessor.getCommand(request.chatId) != Command.START) {
+            responseMessageService.deleteMessage(request.messageId).toMono()
+        } else {
+            doUpdateDialog(request)
         }
 
     private fun processCommand(request: RequestMessageItem): Mono<ResponseMessageItem> =
         userDaoService.findUser(request.chatId)
             .flatMap { userOpt ->
                 val command = Command.get(request.input)
-
                 if (userOpt.isPresent) {
                     when {
                         request.isTextUpdate && command == Command.START ->
@@ -95,37 +87,35 @@ class ProcessorService(
                             responseMessageService.deleteMessage(request.messageId).toMono()
 
                         request.isButtonUpdate ->
-                            doProcessCommand(user = userOpt.get(), command = command ?: Command.MENU)
+                            doProcessCommand(request = request, user = userOpt.get(), command = command ?: Command.MENU)
 
                         else -> throw RuntimeException("unreachable code")
                     }
                 } else {
-                    loginMessageService.init(request)
-                        .flatMap {
-                            doProcessCommand(
-                                user = UserEntity(telegramId = request.chatId),
-                                command = Command.START
-                            )
-                        }
+                    doProcessCommand(
+                        request = request,
+                        user = UserEntity(telegramId = request.chatId),
+                        command = Command.START
+                    )
                 }
             }
 
-    private fun doUpdateDialog(chatId: Long, option: String): Mono<ResponseMessageItem> =
-        dialogProcessor.processOption(telegramId = chatId, option = option)
+    private fun doUpdateDialog(request: RequestMessageItem): Mono<ResponseMessageItem> =
+        dialogProcessor.processOption(request)
             .flatMap { dto ->
-                val state = dto.state!!
-
-                if (state.isFinished == true) {
-                    findExecutor(state.command!!).execute(user = dto.user!!, options = state.options)
+                val state = dto.state
+                if (state.isFinished) {
+                    findExecutor(state.command!!).execute(user = dto.user, options = state.options)
                 } else {
                     state.responseMessageItem!!.toMono()
                 }
             }
 
-    private fun doProcessCommand(user: UserEntity, command: Command): Mono<ResponseMessageItem> =
+    private fun doProcessCommand(request: RequestMessageItem, user: UserEntity, command: Command): Mono<ResponseMessageItem> =
         if (command.isDialog) {
-            dialogProcessor.startDialog(user = user, command = command)
-                .map { it.state?.responseMessageItem!! }
+            dialogProcessor
+                .startDialog(request = request, user = user, command = command)
+                .map { it.state.responseMessageItem!! }
         } else {
             findExecutor(command).execute(user = user)
         }
